lbbgflash: split into its three original translation units
The .sdata2 pool holds three copies of 1.0f, 0.0f and the int-to-double
magic, which only happens when the constants come from separate objects. The
duplicates fall on clean semantic boundaries: screen fades and their renderer
(lbbgflash.c), skeletal transform and two-joint IK (lb_020A.c), and the
LbBf.dat-driven colour animations (lb_0219.c). The per-function data use
agrees: lb_020A.c references only the jobj.h assert strings and owns no data
of its own, while lbbgflash.c and lb_0219.c hold the .bss and .data each of
them uses.

Every function in lbbgflash.c and lb_0219.c now matches; lb_020A.c has one
function left at 98.43%.
