+++ src/melee/ft/ft_0899.c
@@ -5,6 +5,7 @@
 #include "ft/ftparts.h"
 #include "ft/inlines.h"
 #include "lb/lb_00B0.h"
+#include "lb/lb_020A.h"
 #include "lb/lbbgflash.h"
 #include "lb/lbvector.h"
 #include "mp/mplib.h"
