+++ src/melee/lb/lb_0219.h
@@ -1,0 +1,21 @@
+/**
+ * @file lb_0219.h
+ * @brief Data-driven animated screen flashes.
+ */
+
+#ifndef GALE01_0219E4
+#define GALE01_0219E4
+
+#include <platform.h>
+
+#include <baselib/forward.h>
+
+/* 0219E4 */ void fn_800219E4(void*);
+/* 021A10 */ void lbBgFlash_80021A10(f32);
+/* 021A18 */ void lbBgFlash_80021A18(int);
+/* 021B04 */ void fn_80021B04(HSD_GObj*);
+/* 021C1C */ void fn_80021C1C(void);
+/* 021C48 */ void lbBgFlash_80021C48(u32, u32);
+/* 021C80 */ void fn_80021C80(HSD_GObj* gobj);
+
+#endif
